Enable netlink as default collector for vf stats
Add the vfStats package and netlink methods to pull stats for VFs from Netlink.

Netlink is quicker and more widely supported method for collecting VF metrics. It is only supported on kernels 4.4+ and i40e drivers 2.11+. The driver specific stats collection is still implemented to allow these stats to be collected from /sys/fs. This behaviour is no longer the default.