Changing volume types: type: "Directory" to type: "DirectoryOrCreate" to avoid "hostPath type check failed" error.
With this volume type, we can avoid "MountVolume.SetUp failed for volume "kubecgroup" : hostPath type check failed: /sys/fs/cgroup/cpuset/kubepods.slice/ is not a directory"

This way if the volume directory is not present, the directory will be created.
I tested in my deployment and it works well.